chore: apply pr review comments

+++ journal-service/src/main/java/kioke/journal/controller/JournalController.java
@@ -3,6 +3,7 @@ package kioke.journal.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.util.Optional;
+import kioke.commons.exception.security.AccessDeniedException;
 import kioke.commons.http.HttpResponseBody;
 import kioke.journal.constant.Permission;
 import kioke.journal.constant.Role;
@@ -14,7 +15,6 @@ import kioke.journal.dto.response.journal.CreateJournalResponseBodyDto;
 import kioke.journal.dto.response.journal.GetJournalResponseBodyDto;
 import kioke.journal.exception.journal.CannotCreateJournalInArchiveException;
 import kioke.journal.exception.journal.JournalNotFoundException;
-import kioke.journal.exception.permission.NoDeletePermissionsException;
 import kioke.journal.exception.shelf.ShelfNotFoundException;
 import kioke.journal.model.Journal;
 import kioke.journal.model.Shelf;
@@ -69,14 +69,18 @@ public class JournalController {
         .body(HttpResponseBody.success(request, status, data));
   }
 
-  @GetMapping("/{jid}")
+  @GetMapping("/{journalId}")
   public ResponseEntity<HttpResponseBody<GetJournalResponseBodyDto>> getJournal(
-      @AuthenticationPrincipal String uid, @PathVariable String jid, HttpServletRequest request)
+      @AuthenticationPrincipal String userId,
+      @PathVariable String journalId,
+      HttpServletRequest request)
       throws UsernameNotFoundException, JournalNotFoundException {
-    User user = userService.getUserById(uid);
-    Journal journal = journalService.getJournalById(jid);
+    User user = userService.getUserById(userId);
+    Journal journal = journalService.getJournalById(journalId);
 
-    journalRoleService.hasPermission(user, journal, Permission.READ);
+    if (!journalRoleService.hasPermission(user, journal, Permission.READ)) {
+      throw new JournalNotFoundException(journalId);
+    }
 
     HttpStatus status = HttpStatus.OK;
     GetJournalResponseBodyDto data = GetJournalResponseBodyDto.from(journal);
@@ -94,7 +98,7 @@ public class JournalController {
     User user = userService.getUserById(userId);
     Journal journal = journalService.getJournalById(journalId);
 
-    journalRoleService.hasPermission(user, journal, Permission.SHARE);
+    if (!journalRoleService.hasPermission(user, journal, Permission.SHARE)) {}
 
     User invitee = userService.getUserById(requestBodyDto.userId());
     journalRoleService.setRole(invitee, journal, requestBodyDto.role());
@@ -129,11 +133,13 @@ public class JournalController {
   @DeleteMapping("/{jid}")
   public ResponseEntity<HttpResponseBody<Void>> deleteJournal(
       @AuthenticationPrincipal String uid, @PathVariable String jid, HttpServletRequest request)
-      throws UsernameNotFoundException, JournalNotFoundException, NoDeletePermissionsException {
+      throws UsernameNotFoundException, JournalNotFoundException, AccessDeniedException {
     User user = userService.getUserById(uid);
     Journal journal = journalService.getJournalById(jid);
 
-    journalRoleService.hasPermission(user, journal, Permission.DELETE);
+    if (!journalRoleService.hasPermission(user, journal, Permission.DELETE)) {
+      throw new AccessDeniedException();
+    }
 
     Optional<Journal> deletedJournal = journalService.deleteJournal(user, journal);
 
@@ -152,11 +158,13 @@ public class JournalController {
       @PathVariable String journalId,
       @RequestBody @Valid UnshareJournalRequestBodyDto requestBodyDto,
       HttpServletRequest request)
-      throws UsernameNotFoundException, JournalNotFoundException {
+      throws UsernameNotFoundException, JournalNotFoundException, AccessDeniedException {
     User user = userService.getUserById(userId);
     Journal journal = journalService.getJournalById(journalId);
 
-    journalRoleService.hasPermission(user, journal, Permission.SHARE);
+    if (!journalRoleService.hasPermission(user, journal, Permission.SHARE)) {
+      throw new AccessDeniedException();
+    }
 
     User invitee = userService.getUserById(requestBodyDto.userId());
     journalRoleService.deleteRole(invitee, journal);
